Remove OpenSage.Platform
SDL2 is already a platform abstraction - no need for another one.
