+++ src/OpenSage.DataViewer/OpenSage.DataViewer.csproj
@@ -15,6 +15,5 @@
   <ItemGroup>
     <ProjectReference Include="..\OpenSage.Mods.BuiltIn\OpenSage.Mods.BuiltIn.csproj" />
     <ProjectReference Include="..\OpenSage.Game\OpenSage.Game.csproj" />
-    <ProjectReference Include="..\OpenSage.Platform\OpenSage.Platform.csproj" />
   </ItemGroup>
 </Project>
