--- src/OpenSage.Platform/OpenSage.Platform.csproj
@@ -1,18 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
-    <RootNamespace>OpenSage</RootNamespace>
-    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-    <LangVersion>latest</LangVersion>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Veldrid.SDL2" Version="4.2.0-beta1-g25db1e07d8" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\OpenSage.Game\OpenSage.Game.csproj" />
-  </ItemGroup>
-
-</Project>
