+++ src/OpenSage.sln
@@ -29,8 +29,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OpenSage.Mods.Bfme2", "Open
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Mods", "Mods", "{9410748F-ABB1-45AF-98E3-98858A8C0010}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OpenSage.Platform", "OpenSage.Platform\OpenSage.Platform.csproj", "{9F140B62-CA22-46B3-BEFB-AD046275C3ED}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OpenSage.Launcher", "OpenSage.Launcher\OpenSage.Launcher.csproj", "{1B444162-022A-4E69-A9D3-00582CCE5586}"
 EndProject
 Global
@@ -83,10 +81,6 @@ Global
 		{771E6271-64F9-4BB4-9E38-DF4DA0029898}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{771E6271-64F9-4BB4-9E38-DF4DA0029898}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{771E6271-64F9-4BB4-9E38-DF4DA0029898}.Release|Any CPU.Build.0 = Release|Any CPU
-		{9F140B62-CA22-46B3-BEFB-AD046275C3ED}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{9F140B62-CA22-46B3-BEFB-AD046275C3ED}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{9F140B62-CA22-46B3-BEFB-AD046275C3ED}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{9F140B62-CA22-46B3-BEFB-AD046275C3ED}.Release|Any CPU.Build.0 = Release|Any CPU
 		{1B444162-022A-4E69-A9D3-00582CCE5586}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{1B444162-022A-4E69-A9D3-00582CCE5586}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{1B444162-022A-4E69-A9D3-00582CCE5586}.Release|Any CPU.ActiveCfg = Release|Any CPU
